Deployment into RENDER.com

+++ YY - Testing Projects/Andy - Randomideas-full-ANDY/routes/ideas.js
@@ -62,22 +62,36 @@ router.post('/', async (req, res) => {
 // Update idea
 router.put('/:id', async (req, res) => {
   try {
-    const updatedIdea = await Idea.findByIdAndUpdate(
-      req.params.id, 
-      {
-        $set:{
-          text: req.body.text,
-          tag: req.body.tag
+
+    const idea = await Idea.findById(req.params.id);
+
+    // MATCH THE USERNAMES 
+    if (idea.username === req.body.username) { 
+
+      const updatedIdea = await Idea.findByIdAndUpdate(
+        req.params.id, 
+        {
+          $set:{
+            text: req.body.text,
+            tag: req.body.tag
+          }
+        },
+        {
+          new: true
         }
-      },
-      {
-        new: true
-      }
-    );
-    res.json({ 
-      success: true, 
-      data: updatedIdea 
+      );
+      return res.json({ 
+        success: true, 
+        data: updatedIdea 
+      });
+    }
+
+    // USERNAMES DO NOT MATCH
+    res.status(403).json({ 
+      success: false, 
+      data: 'You are not authorized to update this resource'
     });
+
   } catch (error) {
     res.status(500).json({ 
       success: false, 
@@ -90,11 +104,25 @@ router.put('/:id', async (req, res) => {
 // Delete idea
 router.delete('/:id', async (req, res) => {
   try {
-    await Idea.findByIdAndDelete(req.params.id);
-    res.json({ 
-      success: true, 
-      data: {} 
+
+    const idea = await Idea.findById(req.params.id);
+
+    // MATCH THE USERNAMES 
+    if (idea.username === req.body.username) {
+      
+      await Idea.findByIdAndDelete(req.params.id);
+      return res.json({ 
+        success: true, 
+        data: {} 
+      });
+    }
+
+    // USERNAMES DO NOT MATCH
+    res.status(403).json({ 
+      success: false, 
+      data: 'You are not authorized to delete this resource'
     });
+
   } catch (error) {
     res.status(500).json({ 
       success: false,
